Updated Text Classification Android to use Task Library
