+++ lite/examples/text_classification/android/app/build.gradle
@@ -1,12 +1,13 @@
 apply plugin: 'com.android.application'
+apply plugin: 'de.undercouch.download'
 
 android {
-    compileSdkVersion 28
-    buildToolsVersion "29.0.0"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "org.tensorflow.lite.examples.textclassification"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -29,27 +30,21 @@ android {
             includeAndroidResources = true
         }
     }
-
-    flavorDimensions "tfliteInference"
-    productFlavors {
-        // The TFLite inference is built using the TFLite Java interpreter.
-        interpreter {
-            dimension "tfliteInference"
-        }
-        // Default: The TFLite inference is built using the TFLite Task library (high-level API).
-        taskApi {
-            getIsDefault().set(true)
-            dimension "tfliteInference"
-        }
-    }
 }
 
+// Download the pre-trained model from the internet
+project.ext.ASSET_DIR = projectDir.toString() + '/src/main/assets'
+apply from:'download_model.gradle'
+
 dependencies {
-    interpreterImplementation project(":lib_interpreter")
-    taskApiImplementation project(":lib_task_api")
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'org.jetbrains:annotations:15.0'
+
+    //Task Text Library dependency
+    implementation 'org.tensorflow:tensorflow-lite-task-text:0.2.0'
 
     testImplementation 'androidx.test:core:1.2.0'
     testImplementation 'junit:junit:4.12'
