+++ lite/examples/text_classification/android/app/download_model.gradle
