--- lite/examples/text_classification/android/lib_interpreter/build.gradle
@@ -1,54 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'de.undercouch.download'
-
-android {
-    compileSdkVersion 28
-    buildToolsVersion "29.0.0"
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 28
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility = '1.8'
-        targetCompatibility = '1.8'
-    }
-    aaptOptions {
-        noCompress "tflite"
-    }
-    testOptions {
-        unitTests {
-            includeAndroidResources = true
-        }
-    }
-}
-
-// Download the pre-trained model from the internet
-project.ext.ASSET_DIR = projectDir.toString() + '/src/main/assets'
-apply from:'download_model.gradle'
-
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-
-    implementation 'org.tensorflow:tensorflow-lite:2.2.0'
-    implementation 'org.tensorflow:tensorflow-lite-metadata:0.1.0-rc1'
-    implementation 'org.jetbrains:annotations:15.0'
-
-    testImplementation 'androidx.test:core:1.2.0'
-    testImplementation 'junit:junit:4.12'
-    testImplementation 'org.robolectric:robolectric:4.3'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-}
-
-project(':lib_interpreter').tasks.withType(Test) {
-  enabled = false
-}
