--- lite/examples/text_classification/android/lib_task_api/download_model.gradle
@@ -1,7 +1,0 @@
-task downloadModelFile(type: Download) {
-    src 'https://storage.googleapis.com/download.tensorflow.org/models/tflite/text_classification/text_classification_v2.tflite'
-    dest project.ext.ASSET_DIR + '/text_classification.tflite'
-    overwrite true
-}
-
-preBuild.dependsOn downloadModelFile
