+++ lite/examples/text_classification/android/settings.gradle
@@ -1,4 +1,3 @@
 rootProject.name = 'TFLite Text Classification Demo App'
 include ':app'
-include ':lib_task_api'
-include ':lib_interpreter'
+
